Final code

+++ server/src/main/java/com/fittipaldi/game/domain/Player.java
@@ -6,7 +6,7 @@ public class Player {
 
 	private String id;
 	private String name;
-	private float score;
+	private double score;
 
 	public String getId() {
 		return id;
@@ -24,11 +24,11 @@ public class Player {
 		this.name = name;
 	}
 
-	public float getScore() {
+	public double getScore() {
 		return score;
 	}
 
-	public void setScore(float score) {
+	public void setScore(double score) {
 		this.score = score;
 	}
